dpy 2 compatibility, disable message cache

+++ info.py
@@ -37,7 +37,7 @@ class info(commands.Cog):
                 await message.clear_reaction("\U0001f5d1")
                 return False
             users = await reaction[0].users().flatten()
-            if ctx.message.author in users and reaction[0].message.id == message.id:
+            if ctx.message.author in users and reaction[0].message.id == message.id and reaction.emoji == "\U0001f5d1":
                 await message.delete()
                 return True
 
@@ -75,10 +75,9 @@ class info(commands.Cog):
         #step two: calculate stats based on those reactions
         for i in list(jmmmapping.keys()):
             jmmmapping[i]['jmmscore'] = jmmmapping[i]['reactions']-jmmmapping[i]['draobmmjreactions']
+            #don't include users with no stats on the leaderboard
             if not jmmmapping[i]['reactions'] and not jmmmapping[i]['draobmmjreactions']:
                 jmmmapping.pop(i)
-            elif jmmmapping[i]['draobmmjreactions'] and not jmmmapping[i]['reactions']:
-                jmmmapping[i]['positivity'] = 0
             else:
                 jmmmapping[i]['positivity'] = round((jmmmapping[i]['reactions']/(jmmmapping[i]['reactions']+jmmmapping[i]['draobmmjreactions']))*100,1)
         self.delay = time.time()-start    
@@ -275,8 +274,8 @@ class info(commands.Cog):
         if loading and deleted:
             await loading.delete()
 
-def setup(bot):
-    bot.add_cog(info(bot))
+async def setup(bot):
+    await bot.add_cog(info(bot))
 
-def teardown(bot):  
-    bot.remove_cog(info(bot))
+async def teardown(bot):  
+    await bot.remove_cog(info(bot))
